--- defines.h
@@ -1,17 +1,0 @@
-#pragma once
-#define TOOL_VERSION "0.2.4"
-#define AUTHUR "¶¹×ÓDoge"
-#define URL "http://www.douzidoge.cn"
-#define DEVICE_UPDATE_CONF_URL "http://www.douzidoge.cn/deviceupdate/deviceupdateconf"
-#define DEVICE_UPDATE_CONF "deviceupdateconf"
-#define ESPTOOL_NAME L"esptool.exe"
-
-#define RECVPACK_SIZE 2048
-
-#define GET_NUMKEY 1
-#define GET_CHARKEY 2
-#define GET_YESORNO 4
-
-//#define string wstring
-
-#define _AFXDLL
